--- a/benchmarks_test.go
+++ b/benchmarks_test.go
@@ -1,3 +1,5 @@
+// +build go1.7
+
 package is
 
 import (