I've rewrote some more functions from regexps:

- Hexcolor
- RGBColor
- MongoID
- Latitude
- Longitude

Here are benchmarks:

```bash
$ go test -bench=.
testing: warning: no tests to run
BenchmarkOldHexcolor-4    	 3000000	       418 ns/op	       0 B/op	       0 allocs/op
BenchmarkNewHexcolor-4    	50000000	        29.6 ns/op	       0 B/op	       0 allocs/op
BenchmarkOldRGBcolor-4    	 1000000	      1233 ns/op	       0 B/op	       0 allocs/op
BenchmarkNewRGBcolor-4    	 5000000	       356 ns/op	      48 B/op	       1 allocs/op
BenchmarkOldMongoID-4     	 2000000	       712 ns/op	       0 B/op	       0 allocs/op
BenchmarkNewMongoID-4     	10000000	       146 ns/op	       0 B/op	       0 allocs/op
BenchmarkOldLatitude-4    	 3000000	       484 ns/op	       0 B/op	       0 allocs/op
BenchmarkNewLatitude-4    	30000000	        50.2 ns/op	       0 B/op	       0 allocs/op
BenchmarkOldLongitude-4   	 3000000	       576 ns/op	       0 B/op	       0 allocs/op
BenchmarkNewLongitude-4   	30000000	        52.0 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/bbrodriges/bench	17.763s
```

The only thing that bothers me is that little nasty allocation in new version of RGBColor.