@alioygur 2016-12-08
@bbrodriges thanks for your contribution.

I just added "build flag" for support down to go 1.1